Use hidden profile_id in delete form

+++ api/delete.php
@@ -20,9 +20,9 @@ if ($profile['user_id'] != $_SESSION['user_id']) {
     exit();
 }
 
-if (isset($_POST['delete'])) {
+if (isset($_POST['delete']) && isset($_POST['profile_id'])) {
     $stmt = $pdo->prepare('DELETE FROM Profile WHERE profile_id = :pid');
-    $stmt->execute(array(':pid' => $_GET['profile_id']));
+    $stmt->execute(array(':pid' => $_POST['profile_id']));
     $_SESSION['success'] = 'Profile deleted';
     header('Location: index.php');
     exit();
@@ -38,6 +38,7 @@ flashMessages();
 <p>Are you sure you want to delete this profile?</p>
 <form method="POST">
     <input type="hidden" name="delete" value="1" />
+    <input type="hidden" name="profile_id" value="<?php echo htmlentities($profile['profile_id']); ?>" />
     <input type="submit" class="btn btn-danger" value="Delete" />
     <input type="submit" class="btn btn-default" value="Cancel" onclick="location.href='index.php'; return false;" />
 </form>
